requirements.txt
serializer
djangorestframework=3.12.2
